fix: remove unused host dependencies from wasm build
