+++ crates/celox/src/testbench_compile.rs
@@ -1,0 +1,28 @@
+use crate::ir::{AbsoluteAddr, Program};
+use celox_frontend_veryl::VerylTestbenchSource;
+use celox_testbench::TestbenchProgram;
+
+pub(crate) fn project_observability(program: &mut Program, source: &VerylTestbenchSource) {
+    let (sites, read_variables) = celox_frontend_veryl::collect_testbench_observability(source);
+    program.runtime_schema.runtime_event_sites.extend(sites);
+    program.runtime_schema.testbench_read_roots = read_variables
+        .into_iter()
+        .filter_map(|var_id| {
+            program
+                .frontend
+                .root_variable(var_id)
+                .map(|(address, _)| address)
+        })
+        .collect();
+}
+
+pub(crate) fn compile_semantic_testbench(
+    program: &Program,
+    source: &VerylTestbenchSource,
+) -> Option<TestbenchProgram<AbsoluteAddr>> {
+    celox_frontend_veryl::compile_semantic_testbench(
+        &program.frontend,
+        source,
+        program.runtime_schema.runtime_event_sites.len(),
+    )
+}
